core/tc_auth/requirements.txt
httpx
python-jose[cryptography]
itsdangerous
pydantic
